refactor(kustomize): flatten copyFile, fix preflight comment, split stage.go
